refactor: 심자 승인 인원수 카운트 타입을 Int로 변경

+++ services/service-nightstudy/src/main/kotlin/com/b1nd/dodamdodam/nightstudy/application/nightstudy/data/response/NightStudyApprovedCountResponse.kt
@@ -6,7 +6,7 @@ data class NightStudyApprovedCountResponse(
     val total: PeriodCount,
 ) {
     data class PeriodCount(
-        val period1: Long,
-        val period2: Long,
+        val period1: Int,
+        val period2: Int,
     )
 }
